requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.20.0
Django==4.0.4
psycopg2==2.9.3
gunicorn==20.1.0
whitenoise==6.0.0
dj-database-url==0.5.0
vk-api==11.9.8
Telethon==1.24.0